Extract Entry to LiveObjectResource
